rename welcomeTour file to WelcomeTour
