+++ app/client/ui/WelcomeTour.ts
