+++ static/locales/en.client.json
@@ -897,7 +897,7 @@
         "Row ID": "Row ID",
         "SHOW COLUMN": "SHOW COLUMN"
     },
-    "welcomeTour": {
+    "WelcomeTour": {
         "Add New": "Add New",
         "Browse our {{templateLibrary}} to discover what's possible and get inspired.": "Browse our {{templateLibrary}} to discover what's possible and get inspired.",
         "Building up": "Building up",
