+++ static/locales/nb_NO.client.json
@@ -761,7 +761,7 @@
         "Row ID": "Rad-ID",
         "SHOW COLUMN": "Vis kolonne"
     },
-    "welcomeTour": {
+    "WelcomeTour": {
         "Configuring your document": "Oppsett av dokumentet ditt",
         "Customizing columns": "Oppsett av kolonner",
         "Building up": "Oppbygging",
