+++ static/locales/pl.client.json
@@ -892,7 +892,7 @@
         "Row ID": "Identyfikator wiersza",
         "SHOW COLUMN": "POKAŻ KOLUMNĘ"
     },
-    "welcomeTour": {
+    "WelcomeTour": {
         "Add New": "Dodaj nowy",
         "Building up": "Budowanie",
         "Flying higher": "Latanie wyżej",
