+++ static/locales/uk.client.json
@@ -889,7 +889,7 @@
         "Row ID": "ID рядка",
         "SHOW COLUMN": "ПОКАЗАТИ СТОВПЕЦЬ"
     },
-    "welcomeTour": {
+    "WelcomeTour": {
         "Add New": "Додати",
         "Building up": "Створюйте",
         "Configuring your document": "Налаштування вашого документа",
