+++ static/locales/zh_Hans.client.json
@@ -263,7 +263,7 @@
         "Open Snapshot": "打开快照",
         "Snapshots are unavailable.": "快照不可用。"
     },
-    "welcomeTour": {
+    "WelcomeTour": {
         "Share": "分享",
         "Reference": "参考",
         "Flying higher": "飞得更高",
